Ejercicio234

+++ src/main/java/com/josehs/tema07/Ejercicio3.java
@@ -6,12 +6,12 @@ import java.util.Scanner;
 public class Ejercicio3 {
     public static Scanner scanner = new Scanner(System.in);
     public static void main(String[] args) {
-        int nMAlto = Integer.MIN_VALUE;
+        double nMAlto = Integer.MIN_VALUE;
         boolean hayNumeros = false;
 
         System.out.println("Introduce números (introducir letras finaliza el bucle):");
 
-        int numero = pedirNumero();
+        double numero = pedirNumero();
         while (numero == scanner.nextDouble()) {
             hayNumeros = true;
             if (numero > nMAlto) {
@@ -28,10 +28,10 @@ public class Ejercicio3 {
     }
 
     public static int pedirNumero() {
-        int numero = 0;
+        double numero = 0;
         try {
-            if (scanner.hasNextInt()) {
-                numero = scanner.nextInt();
+            if (scanner.hasNextDouble()) {
+                numero = scanner.nextDouble();
             } else {
                 scanner.next();
             }
@@ -39,7 +39,7 @@ public class Ejercicio3 {
             System.out.println("Entrada inválida. Introduce un número válido.");
             scanner.next();
         }
-        return numero;
+        return (int) numero;
     }
 
 }
